Added ssl support with instructions and also cleaned up the naming around vhosts.  There are likely breaking changes with this commit

+++ rdc/redcap-overrides/web/startup-scripts/10-add_var_log_redcap.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
 
 # TO MIMIC OUR INTERNAL SERVERS, MAKE /var/log/redcap A SYM LINK TO /var/log/apache2
-ln -s /var/log/apache2 /var/log/redcap
+if [[ ! -e "/var/log/redcap" ]]; then
+    ln -s /var/log/apache2 /var/log/redcap
+    echo "Linking /var/log/redcap to apache logs"
+fi
